moving flat vector stuffs into a new hnsw package to indicate its usage
