+++ lucene/test-framework/src/java/org/apache/lucene/tests/util/hnsw/HnswTestUtil.java
@@ -23,7 +23,7 @@ import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Deque;
 import java.util.List;
-import org.apache.lucene.codecs.HnswGraphProvider;
+import org.apache.lucene.codecs.hnsw.HnswGraphProvider;
 import org.apache.lucene.codecs.perfield.PerFieldKnnVectorsFormat;
 import org.apache.lucene.index.CodecReader;
 import org.apache.lucene.index.FilterLeafReader;
